9493: Split ConsolidatedCasesCheckboxes into its own file
